requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-filter>=24.0
psycopg2-binary>=2.9
